decomp progress: batch improvements across 35 files
Key fixes:
- DeployCountTracker::TranslateRelativeTargets: precompute ratio before loop (77%→99.6%)
- OverdriveTracker::TranslateRelativeTargets: use u9=i8/playercount + precompute factor (69%→96.8%)
- FileLocalize: add GetGfxMode()==kNewGfx check + fix memcpy offset arithmetic (75%→96.8%)
- Various engine/system functions from prior sessions

+++ src/network/Platform/StringConversion.cpp
@@ -32,7 +32,7 @@ namespace {
                 out++;
                 len--;
             } else {
-                if ((srch - 0xC0) <= 0x1F) {
+                if ((unsigned int)(srch - 0xC0) <= 0x1F) {
                     u8 hi = *in;
                     len--;
                     u8 lo = *((const u8 *)++in);
